correcion en el validador

+++ trunk/seguimiento/item/src/main/java/edu/unlp/informatica/postgrado/seguimiento/item/repository/AbstractRepository.java
@@ -51,8 +51,8 @@ public class AbstractRepository<E, K> {
 		return hibernateTemplate.load(clazz, id);
 	}
 	
-	public int getCount() {
-		return (Integer)hibernateTemplate.findByCriteria(DetachedCriteria.forClass(clazz).setProjection(	Projections.rowCount())).get(0);
+	public long getCount() {
+		return (Long)hibernateTemplate.findByCriteria(DetachedCriteria.forClass(clazz).setProjection(	Projections.rowCount())).get(0);
 	}
 
 }
